fixed decimal method (small syntax error)

+++ src/util/Input.java
@@ -63,7 +63,7 @@ public class Input {
     }
 
     public static double getDouble(){
-        double inputDouble = sc.nextInt();
+        double inputDouble = sc.nextDouble();
         return inputDouble;
     }
